apps/portfolio/serializers.py
from rest_framework import serializers

from apps.portfolio.models import Project, Technology, Screenshot


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = '__all__'


class TechnologySerializer(serializers.ModelSerializer):
    class Meta:
        model = Technology
        fields = '__all__'


class ScreenshotSerializer(serializers.ModelSerializer):
    class Meta:
        model = Screenshot
        fields = ('id', 'project', 'image', 'is_cover', 'caption')
